quora_assessment/requirements.txt
django
djangorestframework
django-crispy-forms